Refactor structure of search page
