+++ src/pages/search/components/NavSearchBar/components/index.js
@@ -1,0 +1,1 @@
+export { default as SearchBar } from './SearchBarV2';
