+++ src/pages/search/components/NavSearchBarDropdownList/components/index.js
@@ -1,0 +1,2 @@
+export { default as SearchBar } from './AlgoliaSearchBar';
+export { default as Hits } from './Hits';
