+++ src/pages/search/components/index.js
@@ -1,0 +1,3 @@
+export { default as MobileSearchBar } from './MobileSearchBar';
+export { default as NavSearchBar } from './NavSearchBar';
+export { default as NavSearchBarDropdownList } from './NavSearchBarDropdownList';
